Release v0.5.0
