+++ Instruments4Music.sln
@@ -5,6 +5,10 @@ VisualStudioVersion = 17.8.34511.84
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Instruments4Music", "Instruments4Music\Instruments4Music.csproj", "{F1CA34D0-2200-4D39-BC45-E8629AD27D82}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MirrorDecor", "MirrorDecorVirtualCamPosFix-main\MirrorDecor.csproj", "{78079984-B8AE-4774-ABAB-AAA1E582E9C7}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BetterSprayPaint", "LCBetterSprayPaint-master\BetterSprayPaint.csproj", "{822A17F7-B413-44DD-8BBB-1475AFCCE008}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,10 +19,14 @@ Global
 		{F1CA34D0-2200-4D39-BC45-E8629AD27D82}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F1CA34D0-2200-4D39-BC45-E8629AD27D82}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F1CA34D0-2200-4D39-BC45-E8629AD27D82}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B59C4274-3A0A-4ED0-BE4C-396C01670908}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B59C4274-3A0A-4ED0-BE4C-396C01670908}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B59C4274-3A0A-4ED0-BE4C-396C01670908}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B59C4274-3A0A-4ED0-BE4C-396C01670908}.Release|Any CPU.Build.0 = Release|Any CPU
+		{78079984-B8AE-4774-ABAB-AAA1E582E9C7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{78079984-B8AE-4774-ABAB-AAA1E582E9C7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{78079984-B8AE-4774-ABAB-AAA1E582E9C7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{78079984-B8AE-4774-ABAB-AAA1E582E9C7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{822A17F7-B413-44DD-8BBB-1475AFCCE008}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{822A17F7-B413-44DD-8BBB-1475AFCCE008}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{822A17F7-B413-44DD-8BBB-1475AFCCE008}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{822A17F7-B413-44DD-8BBB-1475AFCCE008}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
