+++ Instruments4Music/MyPluginInfo.cs
@@ -1,0 +1,9 @@
+namespace Instruments4Music
+{
+    public static class MyPluginInfo
+    {
+        public const string PLUGIN_GUID = "com.bean.Instruments4Music";
+        public const string PLUGIN_NAME = "Instruments4Music";
+        public const string PLUGIN_VERSION = "0.5.0";
+    }
+}
